WIP implement non-functional staticcheck fixes and package consolidations
